Funcioonar Hoteles y Restaurantes con La ApiService

+++ magiacafetera/src/main/java/com/example/magiacafetera/ui/lugares/AdapterLugares.kt
@@ -8,7 +8,6 @@ import com.example.magiacafetera.R
 import com.example.magiacafetera.databinding.CardViewLugaresBinding
 import com.example.magiacafetera.model.DataLugaresItem
 import com.squareup.picasso.Picasso
-import kotlin.math.log
 
 class AdapterLugares(
     private val dataLugares: ArrayList<DataLugaresItem>,
@@ -40,7 +39,7 @@ class AdapterLugares(
         fun render(dataLugares: DataLugaresItem) {
 
             binding.titleTextView.text = dataLugares.title
-            binding.puntuacionRatingBar.rating = dataLugares.puntuacion
+            binding.puntuacionRatingBar.rating = dataLugares.puntuacion.toFloat()
             binding.descripcionTextView.text = dataLugares.descripcion
             Picasso.get().load(dataLugares.image).into(binding.rvImageView)
         }
